+++ src/main/kotlin/com/github/brcosta/cljstuffplugin/actions/EvaluateInlineBaseAction.kt
@@ -11,165 +11,175 @@ import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.CommonDataKeys
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.diagnostic.Logger
-import com.intellij.openapi.editor.*
-import com.intellij.openapi.editor.colors.EditorColorsScheme
+import com.intellij.openapi.editor.Document
+import com.intellij.openapi.editor.Editor
+import com.intellij.openapi.editor.EditorFactory
 import com.intellij.openapi.editor.ex.EditorEx
-import com.intellij.openapi.editor.impl.ComplementaryFontsRegistry
-import com.intellij.openapi.editor.impl.FontInfo
-import com.intellij.openapi.editor.markup.TextAttributes
 import com.intellij.openapi.fileTypes.FileType
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.util.Disposer
+import com.intellij.openapi.util.Computable
 import com.intellij.openapi.util.text.StringUtil
 import com.intellij.ui.EditorTextField
+import com.intellij.ui.awt.RelativePoint
+import com.intellij.util.ConcurrencyUtil
 import cursive.file.ClojureFileType
 import cursive.psi.ClojurePsiElement
 import cursive.repl.actions.ReplAction
-import org.jetbrains.annotations.NotNull
-import java.awt.*
+import java.awt.Component
+import java.awt.Point
+import java.util.concurrent.Executors
+import java.util.concurrent.TimeUnit
+import java.util.concurrent.TimeoutException
+import javax.swing.*
 
-@Suppress("UnstableApiUsage")
+
 open class EvaluateInlineBaseAction(private val formFn: IFn) : AnAction() {
     private val log = Logger.getInstance(AnAction::class.java)
+    private val threadPool = Executors.newCachedThreadPool(
+        ConcurrencyUtil.newNamedThreadFactory("clojure-extras", true, Thread.NORM_PRIORITY)
+    )
 
     override fun actionPerformed(event: AnActionEvent) {
+        println("START OF BASE ACTION  ${Thread.currentThread().name}")
+        val editor = event.getData(CommonDataKeys.EDITOR) ?: return
 
-        event.getData(CommonDataKeys.EDITOR)?.let {
-            (ApplicationManager.getApplication()).invokeAndWait {
-                clearEditorInlays(it)
-                addInlineElement(it, "(*) Evaluating....")
-            }
+        threadPool.submit {
+            println("START OF RUNNABLE THREAD  ${Thread.currentThread().name}")
 
-            (ApplicationManager.getApplication()).invokeLater {
-                val form = this.formFn.invoke(it) as ClojurePsiElement?
-                if (form == null) {
-                    clearEditorInlays(it)
-                    addInlineElement(it, "(x) Could not find form")
-                } else {
-                    processForm(form, it as EditorEx)
+            val form = ApplicationManager.getApplication()
+                .runReadAction(Computable { this.formFn.invoke(editor) as ClojurePsiElement? })
+
+            if (form == null) {
+                ApplicationManager.getApplication().invokeLater { addInlineElement(editor, "(x) Could not find form") }
+            } else {
+
+                val nrepl = NReplClient()
+                ApplicationManager.getApplication()
+                    .invokeLater { addInlineElement(editor, "(*) Evaluating....") { nrepl.interrupt()?.get(); } }
+
+                val result = processForm(form, editor as EditorEx, nrepl)
+                if (nrepl.isConnected) {
+                    nrepl.disconnect()
                 }
+
+                ApplicationManager.getApplication().invokeLater { addInlineElement(editor, result) }
             }
+
         }
+
+        println("END OF BASE ACTION ${Thread.currentThread().name} ")
+
     }
 
+
     private fun processForm(
-        form: ClojurePsiElement, editor: EditorEx
-    ) {
+        form: ClojurePsiElement, editor: EditorEx, nrepl: NReplClient
+    ): String? {
 
-        val stateAtom = ReplAction.replState(editor.project)?.deref() as ILookup?
-        if (stateAtom == null) {
-            addInlineElement(editor, "(x) Repl disconnected")
-        }
+        val stateAtom = ReplAction.replState(editor.project)?.deref() as ILookup? ?: return "(x) Repl disconnected"
 
-        val replState = (stateAtom?.valAt(Keyword.intern("repl-state")) as Atom?)?.deref() as ILookup?
+        val replState = (stateAtom.valAt(Keyword.intern("repl-state")) as Atom?)?.deref() as ILookup?
         val host = replState?.valAt(Keyword.intern("host")) as String?
         val port = replState?.valAt(Keyword.intern("port")) as Long?
         val project = editor.project
 
         if (project != null && host != null && port != null) {
 
-            val namespace = form.ns?.qualifiedName
-            val sessionId = (stateAtom?.valAt(Keyword.intern("session-id"))) as String
+            val namespace =
+                ApplicationManager.getApplication()
+                    .runReadAction(Computable { form.ns?.qualifiedName })
 
-            val nrepl = NReplClient()
+            val text = ApplicationManager.getApplication()
+                .runReadAction(Computable { form.text })
+
+            val sessionId = (stateAtom.valAt(Keyword.intern("session-id"))) as String
             nrepl.connect(host, port.toInt(), sessionId)
 
-            clearEditorInlays(editor)
-
             if (!nrepl.isConnected) {
-                addInlineElement(editor, "(x) Couldn't connect to running session")
+                return "(x) Couldn'' connect to running session"
             } else {
-                val result = nrepl.eval(form.text, namespace).get()
-                if (nrepl.isConnected) {
-                    when {
-                        result["err"] != null -> {
-                            val errorTokens = (result["err"] as String).split("\n").filter { it.isNotEmpty() }
-                            val errorStr = if (errorTokens.size > 1) errorTokens[1] else (errorTokens[0])
-                            log.info("Error: $errorStr")
-                            addInlineElement(editor, "(x) $errorStr")
-                        }
-                        result["value"] != null -> addInlineElement(editor, "=> ${result["value"] as String}")
-                        else -> {
-                            if (result["status"] != null && (result["status"] as ArrayList<*>).size >= 3) {
-                                addInlineElement(editor, "=> ${(result["status"] as ArrayList<*>)[0]}")
+                try {
+                    val result = nrepl.eval(text, namespace).get(30, TimeUnit.SECONDS)
+                    if (nrepl.isConnected) {
+                        when {
+                            result["err"] != null -> {
+                                val errorTokens = (result["err"] as String).split("\n").filter { it.isNotEmpty() }
+                                val errorStr = if (errorTokens.size > 1) errorTokens[1] else (errorTokens[0])
+                                log.info("Error: $errorStr")
+                                return "(x) $errorStr"
+                            }
+                            result["value"] != null -> return "=> ${result["value"] as String}" //addInlineElement(editor, "=> ${result["value"] as String}")
+                            else -> {
+                                if (result["status"] != null && (result["status"] as ArrayList<*>).size >= 3) {
+                                    return "=> ${(result["status"] as ArrayList<*>)[0]}"
+                                }
                             }
                         }
                     }
+                } catch (e: TimeoutException) {
+                    nrepl.interrupt()?.get(5, TimeUnit.SECONDS)
+                    return "(x) Eval timed out"
+                }
+                if (nrepl.isConnected) {
+                    nrepl.disconnect()
                 }
             }
         }
+        return null
     }
 
-    private fun addInlineElement(editor: Editor, text: String) {
-        val document = EditorFactory.getInstance().createDocument(StringUtil.convertLineSeparators(text))
-        val myInput = MyEditorComponent(
-            document, editor.project, ClojureFileType.getInstance(), isViewer = true, oneLineMode = false
-        )
-        HintManager.getInstance().showInformationHint(editor, myInput)
+    private fun addInlineElement(editor: Editor, text: String?, action: Runnable? = null) {
+        if (text != null) {
+
+            val document = EditorFactory.getInstance().createDocument(StringUtil.convertLineSeparators(text))
+            val editorInput = ClojureEditorComponent(
+                document, editor.project, ClojureFileType.getInstance(), isViewer = true, oneLineMode = false
+            )
+            editorInput.border = BorderFactory.createEmptyBorder()
+            val panel = JPanel()
+            panel.layout = BoxLayout(panel, BoxLayout.Y_AXIS)
+            panel.border = BorderFactory.createEmptyBorder()
+            panel.add(editorInput)
+
+            if (action != null) {
+                val btn = JButton("Stop")
+                btn.alignmentX = Component.CENTER_ALIGNMENT
+                btn.addActionListener {
+                    action.run()
+                    HintManager.getInstance().showInformationHint(editor, JLabel("Interrupted"))
+                }
+                panel.add(btn)
+
+                val cursorAbsoluteLocation = editor.visualPositionToXY(editor.caretModel.visualPosition)
+                val editorLocation = editor.component.locationOnScreen
+                val editorContentLocation = editor.contentComponent.locationOnScreen
+                val popupLocation = Point(
+                    editorContentLocation.x + cursorAbsoluteLocation.x,
+                    editorLocation.y + cursorAbsoluteLocation.y - editor.scrollingModel.verticalScrollOffset
+                )
+
+                HintManager.getInstance().showHint(
+                    panel,
+                    RelativePoint.fromScreen(popupLocation),
+                    HintManager.HIDE_BY_OTHER_HINT,
+                    60000
+                )
+            } else {
+                HintManager.getInstance().showInformationHint(editor, panel)
+            }
+
+
+        }
     }
 
-    private fun clearEditorInlays(@NotNull editor: Editor) {
-        editor.inlayModel.getAfterLineEndElementsInRange(
-            0, editor.document.textLength, TextLabelCustomElementRenderer::class.java
-        ).forEach(Disposer::dispose)
-    }
-
-    class MyEditorComponent(
-        document: Document?,
-        project: Project?,
-        fileType: FileType?,
-        isViewer: Boolean,
-        oneLineMode: Boolean
+    class ClojureEditorComponent(
+        document: Document?, project: Project?, fileType: FileType?, isViewer: Boolean, oneLineMode: Boolean
     ) : EditorTextField(document, project, fileType, isViewer, oneLineMode) {
 
         override fun createEditor(): EditorEx {
             val editor = super.createEditor()
             editor.settings.isUseSoftWraps = true
             return editor
-        }
-    }
-
-    class TextLabelCustomElementRenderer(label: String) : EditorCustomElementRenderer {
-        private val label: String
-
-        override fun calcWidthInPixels(inlay: Inlay<*>): Int {
-            val fontInfo: FontInfo = getFontInfo(inlay.editor)
-
-            return fontInfo.fontMetrics().stringWidth(label)
-        }
-
-        override fun paint(inlay: Inlay<*>, g: Graphics, r: Rectangle, textAttributes: TextAttributes) {
-            val editor: Editor = inlay.editor
-            val g2 = g as Graphics2D
-
-            val attributes: TextAttributes = editor.colorsScheme.getAttributes(TEXT_ATTRIBUTES) ?: return
-            val fgColor = attributes.foregroundColor ?: return
-            val fontInfo: FontInfo = getFontInfo(editor)
-            val ascent: Int = editor.ascent
-
-            g2.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON)
-            g2.color = attributes.backgroundColor ?: return
-            g2.fillRoundRect(r.x, r.y + 1, calcWidthInPixels(inlay) + 1, fontInfo.fontMetrics().height + 3, 16, 16)
-            g2.font = Font(fontInfo.font.name, Font.ITALIC, 11)
-            g2.color = fgColor
-
-            g2.drawString(label, r.x + 2, r.y - 1 + ascent)
-        }
-
-        companion object {
-            private val TEXT_ATTRIBUTES = DefaultLanguageHighlighterColors.INLINE_PARAMETER_HINT
-            private fun getFontInfo(editor: Editor): FontInfo {
-                val colorsScheme: EditorColorsScheme = editor.colorsScheme
-                val fontPreferences = colorsScheme.fontPreferences
-                val fontStyle = Font.ITALIC
-                return ComplementaryFontsRegistry.getFontAbleToDisplay(
-                    'a'.toInt(), fontStyle, fontPreferences, FontInfo.getFontRenderContext(editor.contentComponent)
-                )
-            }
-        }
-
-        init {
-            this.label = " $label"
         }
     }