+++ src/main/kotlin/com/github/brcosta/cljstuffplugin/util/AppSettingsConfigurable.kt
@@ -41,6 +41,6 @@ class AppSettingsConfigurable : Configurable {
     }
 
     override fun getDisplayName(): String {
-        return "Clj-Stuff Settings"
+        return "Clojure Extras"
     }
 }